requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
dj-database-url
PyJWT
django-environ
django-filter
django-autoslug
django-countries
Pillow
django-phonenumber-field
phonenumbers
psycopg2-binary
flake8
gunicorn
django-cleanup
beautifulsoup4
whitenoise
requests